Terraform: Day 9, Part 2
max_range_length is a tunable parameter - it's a bit of a cheat so we
don't have to check O(n^2) numbers, which is too slow in terraform. It
would get there eventually without that setting though.

For my input, the range it finds is 17 items long, so 20 is just big
enough.